+++ docs/doxygen/doxygen.dox
@@ -949,7 +949,7 @@ WARN_LOGFILE           =
 # spaces. See also FILE_PATTERNS and EXTENSION_MAPPING
 # Note: If this tag is empty the current directory is searched.
 
-INPUT                  = ../libs/runcleo/ ../libs/superdrops/ ../libs/zarr/ ../libs/observers/
+INPUT                  = ../libs/runcleo/ ../libs/superdrops/ ../libs/zarr/ ../libs/observers/ ../libs/initialise/
 
 # This tag can be used to specify the character encoding of the source files
 # that doxygen parses. Internally doxygen uses the UTF-8 encoding. Doxygen uses
