+++ docs/source/conf.py
@@ -6,7 +6,7 @@ Created Date: Monday 20th November 2023
 Author: Clara Bayley (CB)
 Additional Contributors:
 -----
-Last Modified: Tuesday 9th April 2024
+Last Modified: Friday 19th April 2024
 Modified By: CB
 -----
 License: BSD 3-Clause "New" or "Revised" License
@@ -78,6 +78,7 @@ breathe_projects = {
     "superdrops"       : "../build/doxygen/xml/",
     "zarr"             : "../build/doxygen/xml/",
     "observers"        : "../build/doxygen/xml/",
+    "initialise"       : "../build/doxygen/xml/",
 }
 
 breathe_default_project = 'proj'
