+++ docs/source/cxx/cxx.rst
@@ -21,6 +21,7 @@ a SDM coupled to a Dynamics Solver.
    superdrops/superdropsmod
    zarr/zarrmod
    observers/observersmod
+   initialise/initialisemod
 
 Indices and tables
 ------------------
