+++ docs/source/cxx/initialise/initialconditions.rst
@@ -1,0 +1,8 @@
+InitialConditions
+=================
+
+Header file: ``<libs/initialise/initialconditions.hpp>``
+`[source] <https://github.com/yoctoyotta1024/CLEO/blob/main/libs/initialise/initialconditions.hpp>`_
+
+.. doxygenconcept:: InitialConditions
+   :project: initialise
