--- docs/source/cxx/runcleo/initialconditions.rst
@@ -1,8 +1,0 @@
-InitialConditions
-=================
-
-Header file: ``<libs/runcleo/initialconditions.hpp>``
-`[source] <https://github.com/yoctoyotta1024/CLEO/blob/main/libs/runcleo/initialconditions.hpp>`_
-
-.. doxygenconcept:: InitialConditions
-   :project: runcleo
