+++ libs/initialise/init_all_supers_from_binary.hpp
@@ -32,7 +32,7 @@
 #include <string_view>
 #include <vector>
 
-#include "./initconds.hpp"
+#include "./initialconditions.hpp"
 #include "./optional_config_params.hpp"
 #include "./readbinary.hpp"
 #include "superdrops/superdrop_attrs.hpp"
